requirements.txt
flask
segmentation-models-pytorch
torch
torchvision
numpy
pillow
albumentations
tqdm
